requirements.txt
# Python Dependencies for YOLO Cityscapes Training Project

# Core dependencies
ultralytics>=8.0.0      # YOLOv8 implementation
torch>=2.0.0            # PyTorch deep learning framework
torchvision>=0.15.0     # PyTorch vision utilities

# Image processing
opencv-python>=4.8.0    # Computer vision library
Pillow>=10.0.0          # Python Imaging Library
numpy>=1.24.0           # Numerical computing

# Data processing
tqdm>=4.65.0            # Progress bars
pyyaml>=6.0             # YAML file parsing

# Visualization (optional but recommended)
matplotlib>=3.7.0       # Plotting library
seaborn>=0.12.0         # Statistical visualization

# For dataset preparation
cityscapesscripts>=2.2.0  # Official Cityscapes tools (optional)
roboflow>=1.1.0           # Roboflow dataset download (optional)

# Optional: For faster training
# albumentations>=1.3.0   # Advanced image augmentation
# tensorboard>=2.13.0     # Training visualization
